Fixes edge case in test
- Fixes #5

The `<About>` component should allow users to include a `<p>` tag with
any arbitrary text content. When the content includes a string of "about me",
this matched the selector for the `<h2>` element instead. Updated
tests to search for element using a query selector rather than text content
in order to get around this issue.

Thanks @LNKeat for flagging!